Squashing commits into 1
