+++ tox.ini
@@ -22,7 +22,7 @@ commands =
 
 [testenv:flake8]
 deps = flake8
-commands = flake8 setup.py voluptuous.py tests.py
+commands = flake8 setup.py voluptuous
 
 [testenv:py26]
 basepython = python2.6
