+++ voluptuous/__init__.py
@@ -1,0 +1,11 @@
+try:
+    from schema_builder import *
+    from validators import *
+    from util import *
+except ImportError:
+    from .schema_builder import *
+    from .validators import *
+    from .util import *
+
+__version__ = '0.8.11'
+__author__ = 'tusharmakkar08'
